requirements.txt
streamlit
gspread
pandas
babel
fpdf2[text-shaping]